bgpd: larger one field to be 2 bytes
The `struct attr`'s `encap_tunneltype` should be 2 bytes based on
current code.

+++ bgpd/bgp_attr.h
@@ -213,7 +213,7 @@ struct attr {
 	/* MP Nexthop length */
 	uint8_t mp_nexthop_len;
 
-	uint8_t encap_tunneltype;
+	uint16_t encap_tunneltype;
 
 	/* PMSI tunnel type (RFC 6514). */
 	enum pta_type pmsi_tnl_type;
